Safari-611.1.10/LayoutTests/media/controls/pip-placeholder-without-video-controls-expected.txt

This tests the picture-in-picture placeholder when the video has no controls.

This test only runs in DRT!


EVENT: canplaythrough

Test the picture-in-picture button with valid video

EVENT: webkitpresentationmodechanged
PASS: Should be in pip mode
PASS: Inline placeholder should be visible at this point
PASS: Inline placeholder should have the 'picture-in-picture' class

Testing finished.
